добавили чатрумы и шифрование

+++ MAXrknServer/src/Router.hpp
@@ -14,7 +14,7 @@ class Router {
 public:
 	Router(DBPool&);
 	http::message_generator handle_request(http::request<http::string_body>&&);
-	bool verify_token(const std::string& token, size_t& user_id);
+	bool verify_token(const std::string& token, long long& user_id);
 private:
 	DBPool* dbpool;
 	http::message_generator make_error_responce(const http::request<http::string_body>&, http::status, std::string_view);
